moved files out of components/wizard to be more like other components, properly exporting them
