+++ src/components/Wizard/index.ts
@@ -1,15 +1,14 @@
 import { App } from 'vue'
 import PWizard from './PWizard.vue'
+import PWizardHeaders from './PWizardHeaders.vue'
+import PWizardStep from './PWizardStep.vue'
 import PWizardStepHeader from './PWizardStepHeader.vue'
 
 const install = (app: App): void => {
   app.component('PWizard', PWizard)
+  app.component('PWizardHeaders', PWizardHeaders)
+  app.component('PWizardStep', PWizardStep)
   app.component('PWizardStepHeader', PWizardStepHeader)
 }
-
-export * from './compositions'
-export * from './errors'
-export * from './types'
-export * from './utilities'
 
 export { PWizard, install }
