--- src/components/Wizard/types/index.ts
@@ -1,4 +1,0 @@
-export * from './wizardStep'
-export * from './useWizard'
-export * from './useWizardStep'
-export * from './validation'
