--- src/components/Wizard/types/validation.ts
@@ -1,2 +1,0 @@
-export type WizardStepValidator = () => boolean | Promise<boolean>
-export type ValidationState = { index: number, valid: boolean }
