--- src/components/Wizard/types/wizardStep.ts
@@ -1,7 +1,0 @@
-import { WizardStepValidator } from './validation'
-
-export type WizardStep = {
-  title: string,
-  key?: string,
-  validate?: WizardStepValidator,
-}
