--- src/components/Wizard/utilities/index.ts
@@ -1,1 +1,0 @@
-export * from './getStepKey'
