+++ src/compositions/index.ts
@@ -1,3 +1,4 @@
 export * from './attributes'
 export * from './position'
 export * from './useAdjustedDate'
+export * from './wizard'
