+++ src/compositions/wizard/useWizardStep.ts
@@ -1,8 +1,8 @@
 /* eslint-disable no-redeclare */
 import { inject, InjectionKey, ref, Ref, computed } from 'vue'
-import { WizardNotFound, WizardStepNotFound } from '../errors'
-import { UseWizard, UseWizardStep, WizardStepValidator } from '../types'
 import { useWizardKey } from './useWizard'
+import { WizardNotFound, WizardStepNotFound } from '@/models/wizard'
+import { UseWizard, UseWizardStep, WizardStepValidator } from '@/types/wizard'
 
 export const useWizardStepKey: InjectionKey<UseWizardStep> = Symbol('UseWizardStep')
