+++ src/index.ts
@@ -3,6 +3,7 @@ import '@/styles/index.css'
 export * from '@/components'
 export * from '@/compositions'
 export * from '@/layouts'
+export * from '@/models'
 export * from '@/plugins'
 export * from '@/types'
 export * from '@/utilities'
