+++ src/types/index.ts
@@ -12,3 +12,4 @@ export * from './tables'
 export * from './tabs'
 export * from './units'
 export * from './utilities'
+export * from './wizard'
