+++ src/utilities/index.ts
@@ -9,3 +9,4 @@ export * from './reactivity'
 export * from './seconds'
 export * from './strings'
 export * from './units'
+export * from './wizard'
