+++ src/utilities/wizard.ts
@@ -1,5 +1,5 @@
-import { WizardStep } from '../types'
-import { kebabCase } from '@/utilities'
+import { WizardStep } from '@/types/wizard'
+import { kebabCase } from '@/utilities/strings'
 
 export function getStepKey(step: WizardStep): string {
   return step.key ?? kebabCase(step.title)
